backend/requirements.txt
Django>=4.2,<5.0
asgiref>=3.7

# API
djangorestframework>=3.15,<4.0

# CORS
django-cors-headers>=4.3

# DB
psycopg[binary]>=3.1

# Config
python-dotenv>=1.0

# Static
whitenoise>=6.6

# Worker
celery>=5.3
redis>=5.0

# S3-compatible storage (MinIO/R2)
boto3>=1.34

# Media processing
Pillow>=10.0

# App server
gunicorn>=21.2

# Validation
pydantic>=2.5

# Markdown rendering (PoC1)
mistune>=3.0,<4.0

# Lint
ruff>=0.1
